requirements.txt
django
djangorestframework
mysqlclient
django-cors-headers
djangorestframework-jwt
django_heroku
psycopg2
gunicorn
dj_database_url
